backend/requirements.txt
# Quantum Computing
qiskit
qiskit-machine-learning
qiskit-algorithms
# qiskit-aer

# Deep Learning
torch
torchvision
transformers
timm

# Data Management
datasets
huggingface_hub

# Web Framework
flask
flask-cors

# Networking
pyngrok
requests

# Communications
twilio

# Configuration
python-dotenv

# Data Science
numpy
scikit-learn
scipy
matplotlib
seaborn
pillow

# Utilities
python-dateutil
pytz

# Deployment
gunicorn
